Replace nyulibraries assets with templates, institutions, stylesheets, javascripts
