+++ app/assets/stylesheets/application.scss
@@ -1,7 +1,7 @@
 /*
 *= require font-awesome
 */
-@import "login";
+@import "nyulibraries_stylesheets/login";
 
 .logged_out_text {
   padding: 7px;
