+++ app/assets/stylesheets/application_cu.scss
@@ -1,4 +1,4 @@
 /*
 *= require font-awesome
 */
-@import "login_cu";
+@import "nyulibraries_stylesheets/login_cu";
