+++ app/assets/stylesheets/application_hsl.scss
@@ -1,4 +1,4 @@
 /*
 *= require font-awesome
 */
-@import "login_hsl";
+@import "nyulibraries_stylesheets/login_hsl";
