+++ app/assets/stylesheets/application_ns.scss
@@ -1,4 +1,4 @@
 /*
 *= require font-awesome
 */
-@import "login_ns";
+@import "nyulibraries_stylesheets/login_ns";
