+++ app/assets/stylesheets/application_nysid.scss
@@ -1,4 +1,4 @@
 /*
 *= require font-awesome
 */
-@import "login_nysid";
+@import "nyulibraries_stylesheets/login_nysid";
