+++ app/assets/stylesheets/application_nyuad.scss
@@ -1,4 +1,4 @@
 /*
 *= require font-awesome
 */
-@import "login_nyuad";
+@import "nyulibraries_stylesheets/login_nyuad";
