+++ app/assets/stylesheets/application_nyush.scss
@@ -1,4 +1,4 @@
 /*
 *= require font-awesome
 */
-@import "login_nyush";
+@import "nyulibraries_stylesheets/login_nyush";
